refactor: payment 엔티티 재구성
