+++ src/main/java/roomescape/payment/dto/PaymentRequestDto.java
@@ -1,15 +1,14 @@
-package roomescape.payment.domain.dto;
+package roomescape.payment.dto;
 
 import roomescape.reservation.domain.dto.ReservationWithPaymentDto;
 
-public record PaymentRequestDto(String paymentKey, String orderId, int amount, String paymentType) {
+public record PaymentRequestDto(String paymentKey, String orderId, int amount) {
 
     public static PaymentRequestDto ofReservationWithPaymentDto(ReservationWithPaymentDto requestDto) {
         return new PaymentRequestDto(
                 requestDto.paymentKey(),
                 requestDto.orderId(),
-                requestDto.amount(),
-                requestDto.paymentType()
+                requestDto.amount()
         );
     }
 }
