+++ src/main/java/roomescape/payment/dto/PaymentResponseDto.java
@@ -1,10 +1,10 @@
-package roomescape.payment.domain.dto;
+package roomescape.payment.dto;
 
-import roomescape.payment.domain.TossPayment;
+import roomescape.payment.domain.Payment;
 
 public record PaymentResponseDto(String paymentKey, String orderId, int amount) {
 
-    public static PaymentResponseDto of(TossPayment payment) {
+    public static PaymentResponseDto of(Payment payment) {
         return new PaymentResponseDto(payment.getPaymentKey(), payment.getOrderId(), payment.getAmount());
     }
 }
