+++ src/test/java/roomescape/payment/service/PaymentServiceTest.java
@@ -3,9 +3,9 @@ package roomescape.payment.service;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import roomescape.payment.domain.client.TossRestClient;
-import roomescape.payment.domain.dto.PaymentRequestDto;
-import roomescape.payment.domain.dto.PaymentResponseDto;
+import roomescape.payment.external.TossRestClient;
+import roomescape.payment.dto.PaymentRequestDto;
+import roomescape.payment.dto.PaymentResponseDto;
 import roomescape.payment.exception.InvalidPaymentException;
 import roomescape.payment.exception.PaymentServerException;
 
@@ -13,13 +13,13 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class TossPaymentServiceTest {
+public class PaymentServiceTest {
 
     private final TossRestClient mockRestClient = mock(TossRestClient.class);
 
     @Test
     void 결제_승인_요청시_200_OK() {
-        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000, "NORMAL");
+        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000);
 
         PaymentResponseDto dummyPaymentDto = new PaymentResponseDto("paymentKey", "orderId", 1000);
         when(mockRestClient.confirmPayment(any(PaymentRequestDto.class)))
@@ -34,7 +34,7 @@ public class TossPaymentServiceTest {
 
     @Test
     void 결제_승인_요청시_400에러가_발생하면_InvalidPaymentException_발생() {
-        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000, "NORMAL");
+        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000);
 
         when(mockRestClient.confirmPayment(any(PaymentRequestDto.class)))
                 .thenThrow(new InvalidPaymentException(HttpStatus.BAD_REQUEST));
@@ -48,7 +48,7 @@ public class TossPaymentServiceTest {
 
     @Test
     void 결제_승인_요청시_500에러가_발생하면_PaymentServerException_발생() {
-        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000, "NORMAL");
+        PaymentRequestDto dto = new PaymentRequestDto("paymentKey", "orderId", 1000);
 
         when(mockRestClient.confirmPayment(any(PaymentRequestDto.class)))
                 .thenThrow(new PaymentServerException(HttpStatus.INTERNAL_SERVER_ERROR));
